+++ src/blocks/carousel-content-slides/variations.ts
@@ -1,5 +1,6 @@
 import type { InterpretedAttributes } from "./attributes";
 import type { BlockVariations } from "@atomicsmash/blocks-helpers";
+import { title } from "@wordpress/icons";
 
 /**
  * Variations declared here should be registered in index.tsx
@@ -8,6 +9,7 @@ export const variations = [
 	{
 		name: "heading-and-paragraph",
 		title: "Heading and paragraph",
+		icon: title,
 		innerBlocks: [
 			[
 				"launchpad-blocks/carousel-content-slide",